Move js into one file
